move execution related functions to separate module
