+++ awswrangler/athena/__init__.py
@@ -1,5 +1,11 @@
 """Amazon Athena Module."""
 
+from awswrangler.athena._executions import (  # noqa
+    get_query_execution,
+    stop_query_execution,
+    start_query_execution,
+    wait_query,
+)
 from awswrangler.athena._read import (  # noqa
     get_query_results,
     read_sql_query,
@@ -13,15 +19,11 @@ from awswrangler.athena._utils import (  # noqa
     generate_create_query,
     get_named_query_statement,
     get_query_columns_types,
-    get_query_execution,
     get_query_executions,
     get_work_group,
     list_query_executions,
     repair_table,
     show_create_table,
-    start_query_execution,
-    stop_query_execution,
-    wait_query,
 )
 from awswrangler.athena._write_iceberg import to_iceberg
